trivia: fix failure to insert
using nullable + autoincrement ID was causing generated sqlboiler code
to throw an error complaining about default values while querying out
the result that it had previously inserted. To get around this, generate
an int64 to use as an id. No more than a few hundred players will ever
play this, so it is safe enough.